Refactoring in preparation for firmware modes.
Moving bitbang specific code out of 'main.c'. This way, I can work on
the glitch code without making merging any more difficult.

For now, uncomment the desired mode in int main() (leaving bitbang
here). Later, we can add a 'config.h' or something to make switching
more reasonable.

Tested, working.
